+++ test/router_test.rb
@@ -245,7 +245,6 @@ describe GdsApi::Router do
                        "route_type" => "exact",
                        "handler" => "redirect",
                        "redirect_to" => "/bar",
-                       "redirect_type" => "permanent",
                        "segments_mode" => nil }
         req = WebMock.stub_request(:put, "#{@base_api_url}/routes")
           .with(body: { "route" => route_data }.to_json)
@@ -258,36 +257,17 @@ describe GdsApi::Router do
         assert_requested(req)
       end
 
-      it "should allow creating/updating a temporary redirect route" do
-        route_data = { "incoming_path" => "/foo",
-                       "route_type" => "exact",
-                       "handler" => "redirect",
-                       "redirect_to" => "/bar",
-                       "redirect_type" => "temporary",
-                       "segments_mode" => nil }
-        req = WebMock.stub_request(:put, "#{@base_api_url}/routes")
-          .with(body: { "route" => route_data }.to_json)
-          .to_return(status: 201, body: route_data.to_json, headers: { "Content-type" => "application/json" })
-
-        response = @api.add_redirect_route("/foo", "exact", "/bar", "temporary")
-        assert_equal 201, response.code
-        assert_equal "/bar", response["redirect_to"]
-
-        assert_requested(req)
-      end
-
       it "should allow creating/updating a redirect route which preserves segments" do
         route_data = { "incoming_path" => "/foo",
                        "route_type" => "exact",
                        "handler" => "redirect",
                        "redirect_to" => "/bar",
-                       "redirect_type" => "temporary",
                        "segments_mode" => "preserve" }
         req = WebMock.stub_request(:put, "#{@base_api_url}/routes")
           .with(body: { "route" => route_data }.to_json)
           .to_return(status: 201, body: route_data.to_json, headers: { "Content-type" => "application/json" })
 
-        response = @api.add_redirect_route("/foo", "exact", "/bar", "temporary", segments_mode: "preserve")
+        response = @api.add_redirect_route("/foo", "exact", "/bar", segments_mode: "preserve")
         assert_equal 201, response.code
         assert_equal "/bar", response["redirect_to"]
 
@@ -299,7 +279,6 @@ describe GdsApi::Router do
                        "route_type" => "exact",
                        "handler" => "redirect",
                        "redirect_to" => "bar",
-                       "redirect_type" => "permanent",
                        "segments_mode" => nil }
         response_data = route_data.merge("errors" => { "redirect_to" => "is not a valid URL path" })